Create sweaty-gamer.go I already did the frames.go update below in another PR
Sorry I accidently did the FrameMap Frames.go update for my frames in a different PR please accept that one alongside this one for it to work. I cannot delete the other one. Otherwise I would delete it, and do all of it in one PR, if that would be possible at all.